+++ lib/src/message_listview.dart
@@ -176,10 +176,13 @@ class _MessageListViewState extends State<MessageListView> {
                                 widget.messages[i].user.uid == widget.user.uid
                                     ? MainAxisAlignment.end
                                     : MainAxisAlignment.start,
-                            crossAxisAlignment: CrossAxisAlignment.center,
+                            crossAxisAlignment: CrossAxisAlignment.start,
                             children: <Widget>[
                               Padding(
-                                padding: EdgeInsets.symmetric(horizontal: 10.0),
+                                padding: EdgeInsets.symmetric(
+                                    horizontal:
+                                        MediaQuery.of(context).size.width *
+                                            0.02),
                                 child: ((widget.showAvatarForEverMessage ||
                                             showAvatar) &&
                                         widget.messages[i].user.uid !=
@@ -191,7 +194,9 @@ class _MessageListViewState extends State<MessageListView> {
                                         avatarBuilder: widget.avatarBuilder,
                                       )
                                     : SizedBox(
-                                        width: 40.0,
+                                        width:
+                                            MediaQuery.of(context).size.width *
+                                                0.08,
                                       ),
                               ),
                               GestureDetector(